Added MRT to RadiationSolution
